# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Myanmar University Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: number;
  /** Display name, e.g. Engineering */
  name: string;
  /** Slug used as value in majors.category, e.g. engineering */
  nameEn: string;
  /**
     * Optional badge color hex, e.g. \"#3b82f6\"
     * @nullable
     */
  color?: string | null;
  /** @nullable */
  description?: string | null;
}