# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Myanmar University Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  name: string;
  nameEn: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  description?: string | null;
}